examples/odfe-pipes-trace-app/client/requirements.txt
dash==1.17.0
mysql-connector==2.2.9
opentelemetry-exporter-otlp==0.12b0
opentelemetry-instrumentation-flask==0.12b0
opentelemetry-instrumentation-requests==0.12b0
opentelemetry-sdk==0.12b0
protobuf==3.12.2